// src/@types/Vendedor.d.ts
export interface iVendedor {
  VENDEDOR: number;
  NOME: string;
  ENDERECO: string;
  BAIRRO: string;
  CIDADE: string;
  UF: string;
  CEP: string;
  TELEFONE: string;
  SENHA: string;
  ATUALIZAR: string;
  COMISSAO: number;
  CTPS: string;
  FUNCAO: string;
  ADMISSAO: string;
  DEMISSAO: string;
  SALARIO: number;
  VALE_TRANSPORTE: number;
  NASCIMENTO: string;
  CPF: string | null;
  IDENTIDADE: string | null;
  ESTADO_CIVIL: string;
  PIS: string;
  NACIONALIDADE: string;
  NATURALIDADE: string;
  CONJUGE: string;
  EMAIL: string;
  CELULAR: string;
  ATIVO: string;
  VENDA: string;
  TIPO_VENDEDOR: string;
  CARTAO_NUMERO: string;
  CARTAO_MATRICULA: string;
  META_MARKUP: number;
  META_INDEXADOR: number;
  SETOR: string;
  TABELAS_PERMITIDAS: string | null;
}